refactor: resolve review comments
N
