+++ apiserver/paasng/paasng/core/signals.py
@@ -16,3 +16,7 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
+import django
+
+post_change_app_router = django.dispatch.Signal(providing_args=['application', 'legacy_app', 'migration_process'])
+rollback_change_app_router = django.dispatch.Signal(providing_args=['application', 'legacy_app', 'migration_process'])
